Create GW for hospitalvisit totals
